requirements.txt

streamlit
joblib
pandas
nump
scikit-learn